+++ src/highlighters/substring_highlighter.test.js
@@ -24,8 +24,8 @@ it('highlights a matching partial term in middle of term', () => {
   expect(substringHighlighter('foobarfoo', 'bar')).toEqual(['foo', ['bar'], 'foo']);
 });
 
-it('only highlights this first matching part', () => {
-  expect(substringHighlighter('foobarfoobar', 'bar')).toEqual(['foo', ['bar'], 'foobar']);
+it('highlights all matching parts', () => {
+  expect(substringHighlighter('foobarfoobar', 'bar')).toEqual(['foo', ['bar'], 'foo', ['bar']]);
 });
 
 it('highlights a matching partial term at the end of the term', () => {